requirements.txt
fastapi
numpy
pillow
python-multipart
tensorflow
uvicorn
jinja2



